+++ packages/sz-rust-frontend-codegen/src/generators/vue.rs
@@ -58,6 +58,7 @@ impl<'a> VueComponentGenerator<'a> {
                 source_model: model.name.clone(),
                 source_template: tmpl.to_string(),
                 is_overwritten: false,
+                content,
             });
         }
 
@@ -80,6 +81,7 @@ impl<'a> VueComponentGenerator<'a> {
                     source_model: model.name.clone(),
                     source_template: tmpl.to_string(),
                     is_overwritten: false,
+                    content,
                 });
             }
         }